mod:: API 관련 손보기 - 에러 핸들링 - 스케쥴 추가 로직 파일 변경
