+++ src/components/timetable/StatusBar.tsx
@@ -1,4 +1,5 @@
 import { css } from '@styled-stytem/css'
+import { AxiosError } from 'axios'
 import { useCallback } from 'react'
 
 import { usePostTimetable } from '@/api/hooks/timetable'
@@ -19,11 +20,18 @@ const StatusBar = ({ curSemester, curIndex, setCurIndex }: StatusBarProps) => {
   const curSemesterTimetableLen = curSemester.timetables.length
 
   const handleCreateTimetableBtn = useCallback(() => {
-    createTimetable({
-      timetableName: `timetable ${curSemesterTimetableLen + 1}`,
-      semester: curSemester.semester,
-      year: curSemester.year,
-    })
+    createTimetable(
+      {
+        timetableName: `timetable ${curSemesterTimetableLen + 1}`,
+        semester: curSemester.semester,
+        year: curSemester.year,
+      },
+      {
+        onError: error => {
+          if (error instanceof AxiosError) alert(error.response?.data.message)
+        },
+      },
+    )
   }, [curSemester, createTimetable, curSemesterTimetableLen])
 
   return (
